Refactor 3 files with standard integer types
The following 3 files have been updated, so that integer types in use are now the standard integer types, as opposed to J9 ones.

The 3 files in question are
 - `MemorySubSpaceTarok.cpp`
 - `SchedulingDelegate.cpp`
 - `SchedulingDelegate.hpp`

The types that were changed were:
 - `UDATA` -> `uintptr_t`
 - `U_64` -> `uint64_t`

Additionally, `MM_MemorySubSpaceTarok::checkForRatioExpand()` has been removed, since the new heap sizing logic does not use this function.

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>