Use current pgc cpu time rather than max cpu time for pause time mapping
When pgc pause time is being mapped to a cpu overhead, and the pause time is far below the target, the previous implementation simply returned `dnssExpectedTimeRatioMaximum`.
This is ineffficient, because it effectively prevents eden from shrinking if the overhead is very low (say 1%) and pause time is also low (compared to pause time target).
By returning `_partialGcOverhead` instead when pgc time is very low, it effectively allows cpu overhead alone to dictate eden size.

Additionally, this commit changes the minimum pgc size % from 5% to 1%, and includes a small bug fix when calling `resetPgcTimeStatistics()`

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>