Keep non-eden size the same when eden changes size
This commits primary function, is to make non-eden size remain the same size after eden has resized.
`_heapSizingData.edenRegionChange` is used to communicate the change in eden size to the total heap sizing logic.

Along with this change, the GC cpu overhead being used for total heap resizing, no longer includes PGC (ie. is only looking at GMP cpu overhead).
This is due to the fact that eden and non-eden are increasing/decreasing completely independently of each other.
If PGC overhead is high, it (mostly) has nothing to do with size of non-eden, and so, eden sizing logic looks at that metric.
On the other hand, GMP overhead, is primarily being driven by non-eden size, and so, resizing due to this (instead of PGC overhead as well) is a more sensible approach.

With the fact that the total heap sizing logic no longer considers PGC overhead, the default gc overhead thresholds for eden contraction and expansion, have changed to 2 and 5 respectively (previously 3-9).
The free memory to gc overhead mapping changed very slightly, to account for this change.

Finally, startupPhaseFinished has been removed, since it is now unused

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>